FEATURE: added mrg32 rng

+++ dal/math/vectors.hpp
@@ -133,7 +133,7 @@ namespace Dal {
         Vector_<int> UpTo(int n);
 
         template <class E_>
-        Vector_<E_> XRange(E_ start, E_ finish, int points) {
+        Vector_<E_> XRange(E_ start, E_ finish, size_t points) {
             Vector_<E_> x(points);
             E_ dx = (finish-start)/(points-1);
             for (int i = 0; i < points - 1; ++i)
